requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
google-generativeai==0.5.0
plotly==5.19.0
joblib==1.3.2
python-dotenv==1.0.0
matplotlib==3.8.0